Fix hydration error on page refresh
