+++ src/components/ClientSideIcon.tsx
@@ -1,0 +1,17 @@
+'use client'
+
+import { useEffect, useState } from 'react'
+
+export function ClientSideIcon({ children }: { children: React.ReactNode }) {
+  const [isMounted, setIsMounted] = useState(false)
+
+  useEffect(() => {
+    setIsMounted(true)
+  }, [])
+
+  if (!isMounted) {
+    return null
+  }
+
+  return children
+}
