+++ src/components/FancyNav.jsx
@@ -44,6 +44,8 @@ import {
   CollapsibleContent
 } from './ui/collapsible';
 
+import { ThemeToggleWrapper } from './ThemeToggleWrapper'
+
 const navHierarchy = [
   {
     label: 'My Work',
@@ -160,29 +162,6 @@ const navHierarchy = [
   },
 ];
 
-function ThemeToggle() {
-  let { resolvedTheme, setTheme } = useTheme()
-  let otherTheme = resolvedTheme === 'dark' ? 'light' : 'dark'
-  let [mounted, setMounted] = useState(false)
-
-  useEffect(() => {
-    track("theme-toggle", { theme: otherTheme })
-    setMounted(true)
-  }, [otherTheme])
-
-  return (
-    <button
-      type="button"
-      aria-label={mounted ? `Switch to ${otherTheme} theme` : 'Toggle theme'}
-      className="group rounded-full bg-white/90 px-3 py-2 shadow-lg shadow-zinc-800/5 ring-1 ring-zinc-900/5 backdrop-blur transition dark:bg-zinc-800/90 dark:ring-white/10 dark:hover:ring-white/20"
-      onClick={() => setTheme(otherTheme)}
-    >
-      <SunIcon className="h-6 w-6 fill-zinc-100 stroke-zinc-500 transition group-hover:fill-zinc-200 group-hover:stroke-zinc-700 dark:hidden [@media(prefers-color-scheme:dark)]:fill-teal-50 [@media(prefers-color-scheme:dark)]:stroke-teal-500 [@media(prefers-color-scheme:dark)]:group-hover:fill-teal-50 [@media(prefers-color-scheme:dark)]:group-hover:stroke-teal-600" />
-      <MoonIcon className="hidden h-6 w-6 fill-zinc-700 stroke-zinc-500 transition dark:block [@media(prefers-color-scheme:dark)]:group-hover:stroke-zinc-400 [@media_not_(prefers-color-scheme:dark)]:fill-teal-400/10 [@media_not_(prefers-color-scheme:dark)]:stroke-teal-500" />
-    </button>
-  )
-}
-
 const DesktopNav = ({ navItems }) => (
   <nav className="hidden items-center justify-center gap-6 md:flex">
     <NavigationMenu>
@@ -234,11 +213,11 @@ const DesktopNav = ({ navItems }) => (
           )
         )}
         <NavigationMenuLink className="pl-10 ml-10">
-          <ThemeToggle />
+          <ThemeToggleWrapper />
         </NavigationMenuLink>
       </NavigationMenuList>
     </NavigationMenu>
-  </nav >
+  </nav>
 );
 
 const MobileNav = ({ navItems }) => (
@@ -288,7 +267,7 @@ const MobileNav = ({ navItems }) => (
               )
             )
           )}
-          <ThemeToggle />
+          <ThemeToggleWrapper />
         </nav>
       </div>
     </SheetContent>
